requirements.txt
asgiref==3.9.1
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
crispy-bootstrap5==2025.6
Django==5.2.5
django-crispy-forms==2.4
django-storages==1.14.6
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-storage==3.3.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gunicorn==22.0.0
idna==3.10
mysqlclient==2.2.7
packaging==25.0
pillow==11.3.0
proto-plus==1.26.1
protobuf==4.25.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
requests==2.32.5
rsa==4.9.1
sqlparse==0.5.3
urllib3==2.5.0
django-healthchecks
django-formtools
celery
redis
pandas
numpy==1.26.4
xlsxwriter
openpyxl
django-widget-tweaks
weasyprint
mediapipe==0.10.14
opencv-python-headless==4.8.1.78
tensorflow==2.17.1
tensorflow-hub==0.16.1
tensorflow-probability==0.23.0
